refactor(api): move challenges routes and controller
