+++ api/src/shared/application/challenges/challenge-controller.js
@@ -1,5 +1,5 @@
-import * as challengeRepository from '../../../src/shared/infrastructure/repositories/challenge-repository.js';
-import * as challengeSerializer from '../../infrastructure/serializers/jsonapi/challenge-serializer.js';
+import * as challengeRepository from '../../infrastructure/repositories/challenge-repository.js';
+import * as challengeSerializer from '../../../../lib/infrastructure/serializers/jsonapi/challenge-serializer.js';
 
 const get = async function (request, h, dependencies = { challengeRepository, challengeSerializer }) {
   const challenge = await dependencies.challengeRepository.get(request.params.id);
