+++ api/src/shared/application/challenges/index.js
@@ -1,8 +1,8 @@
 import Joi from 'joi';
 
 import { challengeController } from './challenge-controller.js';
-import { identifiersType } from '../../domain/types/identifiers-type.js';
-import { securityPreHandlers } from '../security-pre-handlers.js';
+import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
+import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
 
 const register = async function (server) {
   server.route([
