+++ api/tests/shared/acceptance/application/challenges/challenge-controller_test.js
@@ -1,5 +1,5 @@
-import { learningContentBuilder, expect, mockLearningContent } from '../../../test-helper.js';
-import { createServer } from '../../../../server.js';
+import { learningContentBuilder, expect, mockLearningContent } from '../../../../test-helper.js';
+import { createServer } from '../../../../../server.js';
 
 describe('Acceptance | API | ChallengeController', function () {
   let server;
